test: Move the service provider and sprout service class tests to match their namespaces
